moved files to seperate directories; +simple pointer checking

+++ WMS/src/sql_manager.c
@@ -1,25 +1,4 @@
-#include "sql_manager.h"
-#include "client.h"
-#define CREATETABLE "CREATE TABLE IF NOT EXISTS 'ITEMS' (\
-		'ID'	    INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT UNIQUE,\
-		'NAME'	    TEXT NOT NULL,\
-		'QUANTITY'	INTEGER NOT NULL DEFAULT 0,\
-		'CLIENT_ID'	INTEGER NOT NULL DEFAULT 0);\
-		 CREATE TABLE IF NOT EXISTS 'CLIENTS' (\
-		'ID'	    INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT UNIQUE,\
-		'NAME'	    TEXT NOT NULL,\
-		'SURNAME'	TEXT NOT NULL);"
-#define INSERTCLIENT "INSERT INTO CLIENTS (ID, NAME, SURNAME) VALUES(, '', '');"
-#define REMOVECLIENT "DELETE FROM CLIENTS WHERE ID=;"
-#define UPDATECLIENT "UPDATE CLIENTS SET NAME='' WHERE ID=;"
-#define SHOWCLIENT "SELECT * FROM CLIENTS WHERE ID=;"
-#define SHOWALLCLIENTS "SELECT * FROM CLIENTS;"
-
-#define INSERTITEM "INSERT INTO ITEMS (ID, NAME, QUANTITY, CLIENT_ID) VALUES(, '', , );"
-#define REMOVEITEM "DELETE FROM ITEMS WHERE ID=;"
-#define SHOWITEM "SELECT * FROM ITEMS WHERE ID=;"
-#define SHOWALLITEMS "SELECT * FROM ITEMS;"
-#define SHOWALLITEMSOFCLIENT "SELECT * FROM ITEMS WHERE CLIENT_ID=;"
+#include "../include/sql_manager.h"
 
 
 void sql_init(sqlite3* db) {
@@ -29,11 +8,9 @@ void sql_init(sqlite3* db) {
 		printf("%s\n", zErrMsg);
 }
 
+
 int default_callback(void* data, int argc, char** argv, char** azColName) {
 	int i;
-	if (argc == 0) {
-		puts("No data to show.");
-	}
 	for (i = 0; i < argc; i++) {
 		printf("%s = %s\n", azColName[i], argv[i] ? argv[i] : "NULL");
 	}
@@ -41,9 +18,7 @@ int default_callback(void* data, int argc, char** argv, char** azColName) {
 	return 0;
 }
 
-int client_creation_callback() {
-	return 0;
-}
+
 
 int client_callback(void* data, int argc, char** argv, char** azColName) {
 	printf("ID: %s\tName: %s\tSurname: %s\n", argv[0], argv[1], argv[2]);
@@ -52,9 +27,15 @@ int client_callback(void* data, int argc, char** argv, char** azColName) {
 
 void sql_addClient(sqlite3 * db, Client * cl) {
 	char* sql = malloc(strlen(INSERTCLIENT) + strlen(cl->NAME) + strlen(cl->SURNAME) + intLen(cl->ID) + 1);
-	int success = 0;
 	char* zErrMsg;
-	success = snprintf(sql, strlen(sql), "INSERT INTO CLIENTS (ID, NAME, SURNAME) VALUES(%d, \'%s\', \'%s\');", cl->ID, cl->NAME, cl->SURNAME);
+	if (sql != NULL)
+	{
+		snprintf(sql, strlen(sql), "INSERT INTO CLIENTS (ID, NAME, SURNAME) VALUES(%d, \'%s\', \'%s\');", cl->ID, cl->NAME, cl->SURNAME);
+	}
+	else {
+		perror("AN ERROR OCCURRED");
+		exit(-1);
+	}
 	sqlite3_exec(db, sql, NULL, 0, &zErrMsg);
 	if (zErrMsg) {
 		puts(zErrMsg);
@@ -65,9 +46,15 @@ void sql_addClient(sqlite3 * db, Client * cl) {
 
 void sql_removeClient(sqlite3 * db, Client * cl) {
 	char* sql = malloc(strlen(REMOVECLIENT) + intLen(cl->ID) + 1);
-	int success = 0;
 	char* zErrMsg;
-	success = snprintf(sql, strlen(sql), "DELETE FROM CLIENTS WHERE ID=%d;", cl->ID);
+	if (sql != NULL)
+	{
+		snprintf(sql, strlen(sql), "DELETE FROM CLIENTS WHERE ID=%d;", cl->ID);
+	}
+	else {
+		perror("AN ERROR OCCURRED");
+		exit(-1);
+	}
 	sqlite3_exec(db, sql, NULL, 0, &zErrMsg);
 	if (zErrMsg) {
 		puts(zErrMsg);
@@ -78,14 +65,20 @@ void sql_removeClient(sqlite3 * db, Client * cl) {
 
 void sql_updateClient(sqlite3 * db, Client * cl) {
 	char* sql = malloc(strlen(UPDATECLIENT) + intLen(cl->ID) + 1);
-	int success = 0;
 	char* zErrMsg, *newName, *newSurname;
 	puts("Please enter new name for the client.");
 	newName = scanString();
 	puts("Please enter new surname for the client");
 	newSurname = scanString();
 
-	success = snprintf(sql, strlen(sql), "UPDATE CLIENTS SET NAME='%s', SURNAME='%s' WHERE ID=%d;", newName, newSurname, cl->ID);
+	if (sql != NULL)
+	{
+		snprintf(sql, strlen(sql), "UPDATE CLIENTS SET NAME='%s', SURNAME='%s' WHERE ID=%d;", newName, newSurname, cl->ID);
+	}
+	else {
+		perror("AN ERROR OCCURRED");
+		exit(-1);
+	}
 	sqlite3_exec(db, sql, NULL, 0, &zErrMsg);
 	if (zErrMsg) {
 		puts(zErrMsg);
@@ -96,9 +89,14 @@ void sql_updateClient(sqlite3 * db, Client * cl) {
 
 void sql_showClient(sqlite3 * db, Client * cl) {
 	char* sql = malloc(strlen(SHOWCLIENT) + intLen(cl->ID) + 1);
-	int success = 0;
 	char* zErrMsg;
-	success = snprintf(sql, strlen(sql), "SELECT * FROM CLIENTS WHERE ID=%d;", cl->ID);
+	if (sql != NULL) {
+		snprintf(sql, strlen(sql), "SELECT * FROM CLIENTS WHERE ID=%d;", cl->ID);
+	}
+	else {
+		perror("AN ERROR OCCURRED");
+		exit(-1);
+	}
 	sqlite3_exec(db, sql, client_callback, 0, &zErrMsg);
 	if (zErrMsg) {
 		puts(zErrMsg);
@@ -108,9 +106,8 @@ void sql_showClient(sqlite3 * db, Client * cl) {
 }
 
 void sql_showAllClients(sqlite3 * db) {
-	int success = 0;
 	char* zErrMsg;
-	sqlite3_exec(db, SHOWALLCLIENTS, default_callback, 0, &zErrMsg);
+	sqlite3_exec(db, SHOWALLCLIENTS, client_callback, 0, &zErrMsg);
 	if (zErrMsg) {
 		puts(zErrMsg);
 	}
@@ -120,9 +117,16 @@ void sql_showAllClients(sqlite3 * db) {
 
 void sql_addItem(sqlite3 * db, Item * it, Client * cl) {
 	char* sql = malloc(strlen(INSERTITEM) + intLen(it->ID) + strlen(it->NAME) + intLen(it->QUANTITY) + intLen(cl->ID) + 1);
-	int success = 0;
 	char* zErrMsg;
-	success = snprintf(sql, strlen(sql), "INSERT INTO ITEMS (ID, NAME, QUANTITY, CLIENT_ID) VALUES(%d, '%s', %d, %d);", it->ID, it->NAME, it->QUANTITY, cl->ID);
+	if (sql != NULL)
+	{
+		snprintf(sql, strlen(sql), "INSERT INTO ITEMS (ID, NAME, QUANTITY, CLIENT_ID) VALUES(%d, '%s', %d, %d);", it->ID, it->NAME, it->QUANTITY, cl->ID);
+	}
+	else
+	{
+		perror("AN ERROR OCCURRED");
+		exit(-1);
+	}
 	sqlite3_exec(db, sql, NULL, 0, &zErrMsg);
 	if (zErrMsg) {
 		puts(zErrMsg);
@@ -134,14 +138,25 @@ void sql_addItem(sqlite3 * db, Item * it, Client * cl) {
 void sql_removeItem(sqlite3 * db, Item * it) {
 	char* sql = malloc(strlen(REMOVEITEM) + intLen(it->ID) + 1);
 	char* zErrMsg;
-	int success;
-	success = snprintf(sql, strlen(sql), "DELETE FROM ITEMS WHERE ID=%d;", it->ID);
+	if (sql != NULL)
+	{
+		snprintf(sql, strlen(sql), "DELETE FROM ITEMS WHERE ID=%d;", it->ID);
+	}
+	else
+	{
+		perror("AN ERROR OCCURRED");
+		exit(-1);
+	}
 	sqlite3_exec(db, sql, NULL, 0, &zErrMsg);
 	if (zErrMsg) {
 		puts(zErrMsg);
 	}
 	sqlite3_free(zErrMsg);
 	free(sql);
+}
+
+void sql_removeAllItems(sqlite3 * db, Client * cl) {
+
 }
 
 void sql_showItem(sqlite3 * db, Item * it) {
@@ -179,3 +194,5 @@ void sql_showAllItemsOfClient(sqlite3 * db, Client * cl) {
 	sqlite3_free(zErrMsg);
 	free(sql);
 }
+
+
